# spec/openapi-dashboard.yml
openapi: "3.0.0"
info:
  version: 0.1.0
  title: E2B Dashboard API
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    AdminApiKeyAuth:
      type: apiKey
      in: header
      name: X-Admin-Token
    AdminJWTAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    # AuthProviderBearerAuth / AuthProviderTeamAuth: B before T in the name
    # so Bearer is validated before Team.
    AuthProviderBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: access_token
    AuthProviderTeamAuth:
      type: apiKey
      in: header
      name: X-Team-ID
  parameters:
...
    clusterID:
      name: clusterID
      in: path
      required: true
      description: Identifier of the cluster.
      schema:
        type: string
        format: uuid
...
  responses:
    "400":
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    "401":
      description: Authentication error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    "500":
      description: Server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
          description: Error code.
        error_code:
          $ref: "#/components/schemas/ErrorCode"
        message:
          type: string
          description: Error message.
    ErrorCode:
      type: string
      description: >-
        Stable machine-readable reason for an API error. New codes may be added
        as other errors adopt this contract. Clients must handle absent or unknown
        codes without interpreting the human-readable message.
      enum:
        - cluster_registration_invalid
        - cluster_registration_conflict
        - cluster_assignment_invalid
        - cluster_assignment_project_not_found
        - cluster_assignment_cluster_not_found
        - cluster_assignment_requires_enterprise
        - cluster_assignment_already_assigned
...
paths:
...
  /v1/management/clusters/{clusterID}/destroy-readiness:
    get:
      operationId: managementClusterDestroyReadiness
      summary: Check cluster destroy readiness
      description: Checks whether the cluster has deletion protection turned on, or more than one team uses this exact cluster, either as an assigned team or as the owner of active templates or snapshots on it. One team's own templates and snapshots do not block this check, because deleting the cluster soft-deletes them. Soft-deleted history does not count. Returns success if the cluster is absent. This read does not change resources or prevent later assignments or template creation.
      tags: [control-plane-management]
      security:
        - AdminJWTAuth: []
      parameters:
        - $ref: "#/components/parameters/clusterID"
      responses:
        "204":
          description: At most one team uses the cluster.
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "409":
          description: The cluster has deletion protection turned on, or more than one team uses it and the other teams and their templates must move off it before destroying the cluster.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/500"
...